third_party/WebKit/LayoutTests/fast/multicol/shadow-breaking.html

<!DOCTYPE html>
<script src="../../resources/check-layout.js"></script>
<p>Test that text-shadow doesn't mess up the line positions.</p>
<div style="position:relative; font:20px/1 Ahem; width:420px; -webkit-columns:2; -webkit-column-gap:20px; orphans:1; widows:1; text-shadow:1px 100px 50px;">
    <span class="testee" data-offset-y="0" data-offset-x="0">line</span><br>
    <span class="testee" data-offset-y="0" data-offset-x="220">line</span><br>
</div>
<p id="result"></p>
<script>
    checkLayout('.testee', document.getElementById('result'));
</script>
